second commit

+++ Entitites/Concrete/Product.cs
@@ -10,7 +10,7 @@ namespace Entitites.Concrete
         public int ProductId { get; set; }
         public int CategoryId { get; set; }
         public string ProductName { get; set; }
-        public int UnitsInStock { get; set; }
+        public short UnitsInStock { get; set; }
         public decimal UnitPrice { get; set; }
     }
 }
